# specs/003-structured-data-query/contracts/api.yaml
openapi: 3.1.0
info:
  title: AIKM 結構化資料查詢 API
  description: 車輛維修知識管理平台結構化資料查詢介面
  version: 1.0.0

servers:
  - url: /api/v1
    description: API v1

tags:
  - name: Query
    description: 自然語言查詢
  - name: Structured
    description: 結構化資料存取
  - name: Dashboard
    description: 儀表板統計
  - name: Export
    description: 資料匯出

paths:
  /query:
    post:
      tags: [Query]
      summary: 統一查詢入口
      description: 接收自然語言查詢，自動識別意圖並路由至對應處理引擎
      operationId: processQuery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
      responses:
        '200':
          description: 查詢成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /structured/vehicles:
    get:
      tags: [Structured]
      summary: 查詢車輛清單
      operationId: listVehicles
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: depot
          in: query
          schema:
            type: string
          description: 機務段篩選
        - name: status
          in: query
          schema:
            type: string
            enum: [active, maintenance, retired]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleListResponse'

  /structured/vehicles/{vehicleCode}:
    get:
      tags: [Structured]
      summary: 查詢單一車輛詳細資料
      operationId: getVehicle
      parameters:
        - name: vehicleCode
          in: path
          required: true
          schema:
            type: string
          example: EMU801
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleDetail'
        '404':
          $ref: '#/components/responses/NotFound'

  /structured/vehicles/{vehicleCode}/faults:
    get:
      tags: [Structured]
      summary: 查詢車輛故障歷程
      operationId: getVehicleFaults
      parameters:
        - name: vehicleCode
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: faultType
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [open, in_progress, resolved]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FaultListResponse'

  /structured/vehicles/{vehicleCode}/maintenance:
    get:
      tags: [Structured]
      summary: 查詢車輛檢修歷程
      operationId: getVehicleMaintenance
      parameters:
        - name: vehicleCode
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: maintenanceType
          in: query
          schema:
            type: string
            enum: [routine, corrective, preventive]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceListResponse'

  /structured/vehicles/{vehicleCode}/costs:
    get:
      tags: [Structured]
      summary: 查詢車輛維修成本
      operationId: getVehicleCosts
      parameters:
        - name: vehicleCode
          in: path
          required: true
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostSummaryResponse'

  /structured/inventory:
    get:
      tags: [Structured]
      summary: 查詢零件庫存
      operationId: listInventory
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: category
          in: query
          schema:
            type: string
        - name: lowStock
          in: query
          schema:
            type: boolean
          description: 只顯示低於安全庫存的項目
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryListResponse'

  /dashboard/summary:
    get:
      tags: [Dashboard]
      summary: 取得儀表板摘要統計
      operationId: getDashboardSummary
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardSummary'

  /dashboard/fault-trends:
    get:
      tags: [Dashboard]
      summary: 取得故障趨勢資料
      operationId: getFaultTrends
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, quarter, year]
            default: month
        - name: depot
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FaultTrendsResponse'

  /dashboard/cost-distribution:
    get:
      tags: [Dashboard]
      summary: 取得維修成本分布
      operationId: getCostDistribution
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [depot, vehicle_type, cost_type]
            default: depot
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostDistributionResponse'

  /export:
    post:
      tags: [Export]
      summary: 匯出資料
      operationId: exportData
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
      responses:
        '200':
          description: 檔案下載
          content:
            text/csv:
              schema:
                type: string
                format: binary
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'

components:
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

  schemas:
    QueryRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: 自然語言查詢
          example: "查詢 EMU801 的故障歷程"
        context:
          type: object
          description: 對話上下文

    QueryResponse:
      type: object
      properties:
        intent:
          type: string
          enum: [knowledge_query, structured_query, hybrid_query, clarification_needed]
        answer:
          type: string
          description: 文字回覆
        data:
          $ref: '#/components/schemas/StructuredData'
        sources:
          type: array
          items:
            $ref: '#/components/schemas/Source'

    StructuredData:
      type: object
      properties:
        type:
          type: string
          enum: [table, card, chart]
        title:
          type: string
        columns:
          type: array
          items:
            $ref: '#/components/schemas/Column'
        rows:
          type: array
          items:
            type: object
        total:
          type: integer
        hasMore:
          type: boolean

    Column:
      type: object
      properties:
        key:
          type: string
        label:
          type: string
        type:
          type: string
          enum: [string, number, date, status]

    Source:
      type: object
      properties:
        type:
          type: string
          enum: [document, database]
        name:
          type: string
        relevance:
          type: number

    Vehicle:
      type: object
      properties:
        id:
          type: string
          format: uuid
        vehicleCode:
          type: string
          example: EMU801
        vehicleType:
          type: string
        status:
          type: string
          enum: [active, maintenance, retired]
        depot:
          type: string
        lastMaintenanceDate:
          type: string
          format: date

    VehicleDetail:
      allOf:
        - $ref: '#/components/schemas/Vehicle'
        - type: object
          properties:
            manufacturer:
              type: string
            manufactureYear:
              type: integer
            recentFaults:
              type: array
              items:
                $ref: '#/components/schemas/FaultRecord'
            recentMaintenance:
              type: array
              items:
                $ref: '#/components/schemas/MaintenanceRecord'

    VehicleListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Vehicle'
        pagination:
          $ref: '#/components/schemas/Pagination'

    FaultRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        faultCode:
          type: string
        faultDate:
          type: string
          format: date-time
        faultType:
          type: string
        severity:
          type: string
          enum: [critical, major, minor]
        status:
          type: string
          enum: [open, in_progress, resolved]
        description:
          type: string

    FaultListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/FaultRecord'
        pagination:
          $ref: '#/components/schemas/Pagination'

    MaintenanceRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        maintenanceCode:
          type: string
        maintenanceDate:
          type: string
          format: date
        maintenanceType:
          type: string
          enum: [routine, corrective, preventive]
        technicianName:
          type: string
        status:
          type: string
          enum: [scheduled, in_progress, completed]
        workHours:
          type: number

    MaintenanceListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/MaintenanceRecord'
        pagination:
          $ref: '#/components/schemas/Pagination'

    CostSummaryResponse:
      type: object
      properties:
        vehicleCode:
          type: string
        totalCost:
          type: number
        laborCost:
          type: number
        materialCost:
          type: number
        externalCost:
          type: number
        breakdown:
          type: array
          items:
            $ref: '#/components/schemas/CostBreakdown'

    CostBreakdown:
      type: object
      properties:
        period:
          type: string
        laborCost:
          type: number
        materialCost:
          type: number
        totalCost:
          type: number

    InventoryItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        partCode:
          type: string
        partName:
          type: string
        category:
          type: string
        currentStock:
          type: integer
        safetyStock:
          type: integer
        unit:
          type: string
        location:
          type: string
        isLowStock:
          type: boolean

    InventoryListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/InventoryItem'
        pagination:
          $ref: '#/components/schemas/Pagination'

    DashboardSummary:
      type: object
      properties:
        totalVehicles:
          type: integer
        activeVehicles:
          type: integer
        openFaults:
          type: integer
        criticalFaults:
          type: integer
        scheduledMaintenance:
          type: integer
        lowStockItems:
          type: integer
        monthlyMaintenanceCost:
          type: number

    FaultTrendsResponse:
      type: object
      properties:
        period:
          type: string
        data:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              count:
                type: integer
              byType:
                type: object
                additionalProperties:
                  type: integer

    CostDistributionResponse:
      type: object
      properties:
        groupBy:
          type: string
        data:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              laborCost:
                type: number
              materialCost:
                type: number
              totalCost:
                type: number

    ExportRequest:
      type: object
      required:
        - table
        - format
      properties:
        table:
          type: string
          enum: [vehicles, fault_records, maintenance_records, usage_records, parts_used, cost_records, parts_inventory]
        format:
          type: string
          enum: [csv, xlsx]
        filters:
          type: object
          additionalProperties: true
        columns:
          type: array
          items:
            type: string
          description: 要匯出的欄位，若為空則匯出全部

    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer

  responses:
    BadRequest:
      description: 請求參數錯誤
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: array
                items:
                  type: string

    Unauthorized:
      description: 未授權
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string

    NotFound:
      description: 資源不存在
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - BearerAuth: []